move validateIndices to helpers
